pyproject.toml
[tool.poetry]
name = "project"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
djangorestframework = "^3.12.4"
pandas = "^1.3.3"
Django = "^3.2.7"
django-cors-headers = "^3.9.0"
django-environ = "^0.7.0"
django-filter = "^21.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"